part4.21, step10 (add request delete blogs only by auth user)

+++ part4/bloglist/controllers/blogs.js
@@ -31,7 +31,7 @@ blogsRouter.post('/', async (request, response, next) => {
   try {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    if(!request.token || !decodedToken.id) {
+    if (!request.token || !decodedToken.id) {
       // 401 unauthorized
       return response.status(401).json({
         error: 'token missing or invalid'
@@ -83,6 +83,28 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!blog)  {
+      return response.status(400).json({
+        error: 'blog not found'
+      })
+    }
+
+    if (!request.token || !decodedToken.id) {
+      // 401 unauthorized
+      return response.status(401).json({
+        error: 'token missing or invalid'
+      })
+    }
+
+    if (!(blog.user.toString() === decodedToken.id.toString())) {
+      return response.status(401).json({
+        error: 'not authorised'
+      })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end() // 204 No Content
   }
